+++ config/initializers/i18n.rb
@@ -1,0 +1,1 @@
+I18n.available_locales = [:en, :ca, :es, :fr, :ar, :de, :ru, :pt, :vi, :tr]