+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 * Add Vue.js
+* Add search results map
+* Display infowindow on maps
 * Add social sharing buttons
-* Add Turkish and Vietnamese
+* Add Turkish and Vietnamese translations
 * Fix Rails 5.1.0 issue with seed task (#29)
 * Support Rails 5.1.0