+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 * Theme can be changed via query string
+
+
+* Add Vue.js
+* Add social sharing buttons
+* Add Turkish and Vietnamese
 * Fix Rails 5.1.0 issue with seed task (#29)
 * Support Rails 5.1.0